api/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
fastf1>=3.4.0
httpx>=0.26.0
sqlalchemy>=2.0.25
pydantic>=2.6.0
python-dotenv>=1.0.1
alembic>=1.13.0
python-multipart>=0.0.9
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0
requests>=2.31.0
joblib>=1.3.0
scipy>=1.12.0
xgboost>=2.0.0